requirements.txt
apache-airflow[postgres,celery,redis]==2.3.2
apache-airflow-providers-apache-spark==4.0.0
apache-airflow-providers-http==2.1.2
pytest==8.4.1